backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
markdown>=3.5.0
watchdog>=3.0.0
